Spark-Connector 1.6.1:
支持 Stage方式写入
支持写入限流
支持设置分区方式读取，目标可以是内表、外表或 view，读取可以按照shard、分区表、字段范围分片
Arrow 方式读取支持 jsonb 类型

+++ hologres-connector-spark-base/src/main/scala/com/alibaba/hologres/spark/sink/FieldWriter.scala
@@ -65,7 +65,7 @@ class StringFieldWriter(removeU0000: Boolean) extends FieldWriter {
 
 class DateFieldWriter extends FieldWriter {
   override def writeValue(row: InternalRow, idx: Int): Date = {
-    Date.valueOf(LocalDate.ofEpochDay(row.getLong(idx)))
+    Date.valueOf(LocalDate.ofEpochDay(row.getInt(idx)))
   }
 }
